feat(challenge): Move useCounter hook #162
