+++ react-advanced/src/customHooks/hooks/useCounter.tsx
@@ -1,0 +1,13 @@
+import { useEffect, useState } from "react";
+
+export const useCounter = () => {
+  const [count, setCount] = useState(0);
+
+  useEffect(() => {
+    const id = setInterval(() => {
+      setCount((c) => c + 1);
+    }, 1000);
+    return () => clearInterval(id);
+  }, []);
+  return count;
+};
